airflow/requirements.txt
apache-airflow-providers-snowflake
apache-airflow-providers-http
snowflake-connector-python
requests